refactor(trading): relocate isBuyState to OfferSignal and breakeven helper to OfferRecord
