+++ src/test/java/com/flipsmart/BreakevenRelistDetectionTest.java
@@ -1,5 +1,5 @@
 package com.flipsmart;
-import com.flipsmart.domain.offer.TrackedOffer;
+import com.flipsmart.domain.offer.OfferRecord;
 
 import org.junit.Test;
 import static org.junit.Assert.assertFalse;
@@ -10,25 +10,25 @@ public class BreakevenRelistDetectionTest
 	@Test
 	public void advancesWhenRelistAtAdvisedPriceWithinTolerance()
 	{
-		assertTrue(TrackedOffer.shouldAdvanceToBreakevenRelist(true, 1048, 1050));
-		assertTrue(TrackedOffer.shouldAdvanceToBreakevenRelist(true, 1050, 1050));
+		assertTrue(OfferRecord.shouldAdvanceToBreakevenRelist(true, 1048, 1050));
+		assertTrue(OfferRecord.shouldAdvanceToBreakevenRelist(true, 1050, 1050));
 	}
 
 	@Test
 	public void doesNotAdvanceWhenBreakevenNotAccepted()
 	{
-		assertFalse(TrackedOffer.shouldAdvanceToBreakevenRelist(false, 1050, 1050));
+		assertFalse(OfferRecord.shouldAdvanceToBreakevenRelist(false, 1050, 1050));
 	}
 
 	@Test
 	public void doesNotAdvanceWhenRelistPriceFarFromAdvised()
 	{
-		assertFalse(TrackedOffer.shouldAdvanceToBreakevenRelist(true, 1200, 1050));
+		assertFalse(OfferRecord.shouldAdvanceToBreakevenRelist(true, 1200, 1050));
 	}
 
 	@Test
 	public void doesNotAdvanceWithoutAdvisedPrice()
 	{
-		assertFalse(TrackedOffer.shouldAdvanceToBreakevenRelist(true, 1050, 0));
+		assertFalse(OfferRecord.shouldAdvanceToBreakevenRelist(true, 1050, 0));
 	}
 }
